chore: moved TestCosmos into tests/netwrork.
