+++ app/tests/tradeFlow.spec.ts
@@ -6,7 +6,7 @@ import { jest } from '@jest/globals'
 import offers from './fixtures/offers.json'
 import codeIds from './fixtures/codeIds.json'
 import { createHubUpdateConfigMsg } from './utils'
-import { TestCosmosChain } from './TestCosmosChain'
+import { TestCosmosChain } from './network/TestCosmosChain'
 import { DefaultError } from '~/network/chain-error'
 import type { GetOffer, PostOffer } from '~/types/components.interface'
 import { TradeState } from '~/types/components.interface'
